requirements.txt
fastapi
uvicorn
torch
transformers
scikit-learn
pandas
streamlit
requests
python-dotenv
accelerate